+++ agent/CHANGELOG.rst
@@ -7,6 +7,7 @@ This file keeps track of all notable changes to license-manager-agent
 Unreleased
 ----------
 * Updated configuration row to use new feature quantities format
+* Added OLicense license server support
 
 2.2.15 -- 2022-10-26
 --------------------